refactor(arch): split the `dummy` arch into multiple submodule files
Split the `dummy` arch to make it more manageable when it keeps growing.
