+++ src/riot-rs-embassy/src/arch/dummy/hwrng.rs
@@ -1,0 +1,5 @@
+use crate::arch;
+
+pub fn construct_rng(_peripherals: &mut arch::OptionalPeripherals) {
+    unimplemented!();
+}
